% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ocr_error.R
\name{do_ocr_replacement}
\alias{do_ocr_replacement}
\title{TBC.}
\usage{
do_ocr_replacement(s, where, orgpat, newpat)
}
\arguments{
\item{string}{A string.}
}
\value{
It returns the \code{string} with a randomly assgined OCR error following
   rules extracted in the ocr_rules dataset. It also comes with the change log of the
   transformation.
}
\description{
\code{do_ocr_replacement} randomly assign a Optical Character Recognition (OCR)
    error to a string. This function was converted from the Python code in Febrl (
    developed by Agus Pudjijono in 2008, refers to reference \url{https://link.springer.com/chapter/10.1007/978-3-642-01307-2_47}.
}
\examples{
do_ocr_replacement('how are you?')

}
