% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_typo_error.R
\name{do_typo_replacement}
\alias{do_typo_replacement}
\title{TBC.}
\usage{
do_typo_replacement(s)
}
\arguments{
\item{string}{A string.}
}
\value{
It returns the \code{string} with a randomly assgined typographic error following
   rules extracted in the typo_rules. It also comes with the change log of the
   transformation.
}
\description{
\code{do_typo_replacement} randomly assign a typographic error to a string.
    This function was converted from the Python code in Febrl (developed by
    Agus Pudjijono in 2008, refers to reference \url{https://link.springer.com/chapter/10.1007/978-3-642-01307-2_47}.
}
\examples{
do_typo_replacement('how are you?')

}
