% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{extract_address}
\alias{extract_address}
\title{Extract addresses.}
\usage{
extract_address(n = 30000, postcode = NA)
}
\arguments{
\item{n}{A number.}

\item{postcode}{A string.}
}
\value{
The output is \code{n} addresses in the form of a data framework
  with \code{n} observations with 5 variables:
  1) postcode of the UK address, 2) country, 3) primary_care_trust
  4) longitude of the address and 5) latitude of the address.
  If \code{postcode} is given, the return addresses are addresses
  having the same outward postcode
}
\description{
\code{extract_address} extract addresses using \code{\link{get_address}}.
}
\examples{
extract_address()
extract_address(100)
extract_address(100, postcode = 'w3')
}
