% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{gen_address}
\alias{gen_address}
\title{Generate an address.}
\usage{
gen_address()
}
\value{
The output is a data frame with 1 observation of 5 variables:
  \enumerate{
  \item postcode of the UK address,
  \item country,
  \item primary_care_trust,
  \item longitude of the address,
  \item latitude of the address.
  }
}
\description{
\code{gen_address} randomly return a UK address out of 30,000 UK addresses.
The UK addresses were extracted from \code{\link{extract_address}}.
}
\examples{
gen_address()
}
